requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.4
django-leaflet==0.29.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
docopt==0.6.2
drf-yasg==1.21.7
GDAL @ file:///C:/Users/Ali/pipwin/GDAL-3.4.3-cp310-cp310-win_amd64.whl#sha256=9893da5532a6c5dadc1fc3bc402ae29d161b7cf190d778f129da84d0e50baa29
idna==3.6
inflection==0.5.1
Js2Py==0.74
packaging==24.0
pipwin==0.5.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyjsparser==2.7.1
PyJWT==2.8.0
PyPrind==2.11.3
pySmartDL==1.3.4
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.1
